Renamed `interrupts` to `idt` and moved to `sys` module.
